+++ src/navigation/components/PactModal.tsx
@@ -1,4 +1,3 @@
-import { useMemo, useState, useEffect } from 'react'
 import { View, ScrollView, Alert } from 'react-native'
 import { Navigation } from 'react-native-navigation'
 
@@ -68,9 +67,6 @@ const Content = () => {
 const Footer = ({ componentId }: { componentId: string }) => {
   const theme = useTheme()
   const isAgreePact = useSettingValue('common.isAgreePact')
-  // const checkUpdate = useDispatch('common', 'checkUpdate')
-  const [time, setTime] = useState(20)
-
   const handleRejct = () => {
     exitApp()
     // Navigation.dismissOverlay(componentId)
@@ -96,38 +92,6 @@ const Footer = ({ componentId }: { componentId: string }) => {
       }, 2e3)
     }
   }
-
-
-  const confirmBtn = useMemo(() => {
-    if (isAgreePact) return { disabled: false, text: '关闭' }
-    return time ? { disabled: true, text: `接受（${time}）` } : { disabled: false, text: '接受' }
-  }, [isAgreePact, time])
-
-  useEffect(() => {
-    if (isAgreePact) return
-    const timeoutTools = {
-      timeout: null as NodeJS.Timeout | null,
-      start() {
-        this.timeout = setTimeout(() => {
-          setTime(time => {
-            time--
-            if (time > 0) this.start()
-            return time
-          })
-        }, 1000)
-      },
-      clear() {
-        if (!this.timeout) return
-        clearTimeout(this.timeout)
-      },
-    }
-    timeoutTools.start()
-    return () => {
-      timeoutTools.clear()
-    }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
-
   return (
     <>
       {
@@ -147,8 +111,8 @@ const Footer = ({ componentId }: { componentId: string }) => {
                 </Button>
               )
         }
-        <Button disabled={confirmBtn.disabled} style={{ ...styles.btn, backgroundColor: theme['c-button-background'] }} onPress={handleConfirm}>
-          <Text color={theme['c-button-font']}>{confirmBtn.text}</Text>
+        <Button style={{ ...styles.btn, backgroundColor: theme['c-button-background'] }} onPress={handleConfirm}>
+          <Text color={theme['c-button-font']}>{isAgreePact ? '关闭' : '接受'}</Text>
         </Button>
       </View>
     </>
@@ -222,4 +186,3 @@ const styles = createStyle({
 })
 
 export default PactModal
-